+++ utils/constants.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-APP_VERSION = '2.2.11'
+APP_VERSION = '2.2.12'
 APP_RELEASE_CANDITATE = False
 DICTIONARY_VERSION = 4